refactor(frontend): streamline Run detail panels
- Extract execution identifiers into a collapsible RunDiagnostics disclosure
- Add log toolbar with wrap toggle and per-stream download
- Tighten artifact table with icons and compact download action
- Keep UnderlineNav tab semantics for Summary/Logs/Artifacts
